+++ src/components/ErrorScreen/index.ts
@@ -1,0 +1,5 @@
+import { ErrorScreen } from './ErrorScreen'
+/** @deprecated MessageScreen は非推奨です。ErrorScreen を使ってください。 */
+const MessageScreen = ErrorScreen
+
+export { MessageScreen, ErrorScreen }
