--- src/components/MessageScreen/index.ts
@@ -1,1 +1,0 @@
-export { MessageScreen } from './MessageScreen'
