+++ src/index.ts
@@ -86,7 +86,7 @@ export { InformationPanel } from './components/InformationPanel'
 export { RightFixedNote } from './components/RightFixedNote'
 export { Tooltip } from './components/Tooltip'
 export { BottomFixedArea } from './components/BottomFixedArea'
-export { MessageScreen } from './components/MessageScreen'
+export { ErrorScreen, MessageScreen } from './components/ErrorScreen'
 export { Calendar } from './components/Calendar'
 export { DatePicker } from './components/DatePicker'
 export { SegmentedControl, SegmentedControlOption } from './components/SegmentedControl'
